requirements.txt
Flask==0.12
requests==2.19.0
PyYAML==5.1
pytest==7.0.0
